fix(7800): correct MARIA CTRL bit map — games now boot past the black screen
Native 7800 games booted to black. Tracing Asteroids showed why: the cart runs
its own reset code, then spins at $D06F-$D076 waiting for zero-page $76, which
only its NMI handler ($D2FC) advances — and that NMI is MARIA's DLI. The DLI
never fired, so $76 stayed frozen and the wait never ended. (The BIOS was never
in the path — that earlier theory was wrong; carts boot straight from their own
reset vector.)

Root cause: the MARIA CTRL register ($3C) bit map was wrong on three bits.
Checked against the MiSTer RTL (control.sv / maria.sv), real MARIA is:
DM (DMA mode) = bits 6:5, colour-kill = bit 7, Kangaroo = bit 2. The crate had
DMA-enable = bit 7, colour-kill = bit 6, Kangaroo = bit 1. So a game enabling DMA
with DM=10 (bit 6) read to us as "DMA off, colour-kill on": MARIA never walked the
display list, never raised the DLI, and the game hung. Correcting the three bit
positions makes DMA run; $76 advances and the frame renders (Asteroids goes from a
uniform black frame to ~18 colours / ~5k non-background pixels).

Enabling DMA exposed two latent issues in code that had never executed on real
content, fixed here so the path is robust:
  - dma_cycles / dma_budget overflowed u8 once a populated display list was
    walked — widened both to u16;
  - a malformed display list could walk forever — bounded the walk by a per-line
    DMA-cycle cap (real MARIA's DMA aborts at end of line).

The cart_boot smoke now asserts a rendered frame (>=2 colours, >=500
non-background px), so a regression back to the black screen fails the test.
Pixel-accurate DL-entry decoding (byte roles / width encoding) is follow-on
fidelity work, tracked in outstanding-work; the cap is a safety bound, not a
correctness fix.

+++ crates/machine-atari-7800/src/lib.rs
@@ -99,7 +99,7 @@ pub struct Atari7800 {
     master_clock: u64,
     clocks_per_frame: u64,
     frame_count: u64,
-    dma_budget: u8,
+    dma_budget: u16,
     line_cycle: u16,
 }
 
@@ -155,7 +155,7 @@ impl Atari7800 {
 
         if self.master_clock.is_multiple_of(2) {
             self.line_cycle += 1;
-            if self.line_cycle > u16::from(self.dma_budget) && !self.maria.wsync_halt() {
+            if self.line_cycle > self.dma_budget && !self.maria.wsync_halt() {
                 self.cpu.tick();
                 if self.cpu.rw {
                     self.cpu.data_in = self.mem_read(self.cpu.addr);
